feat: 프론트엔드 username 유효성 검사 추가
